refactor: rearrange classes and simplify json implicits
