+++ app/controllers/api/v4/rest/MetaController.scala
@@ -1,12 +1,10 @@
 package controllers.api.v4.rest
 
 import javax.inject._
-import models.Entities._
+import models.entities.TooComplexQueryError
+import models.entities.TooComplexQueryError._
 import models.{Backend, GQLSchema}
-import models.entities._
-import models.entities.Configuration.JSONImplicits._
-import models.entities.HealthCheck.JSONImplicits._
-import play.api.Logger
+import play.api.Logging
 import play.api.libs.json._
 import play.api.mvc._
 import sangria.macros._
@@ -19,8 +17,7 @@ import scala.concurrent.ExecutionContext
 
 @Singleton
 class MetaController @Inject()(implicit ec: ExecutionContext, backend: Backend, cc: ControllerComponents)
-  extends AbstractController(cc) {
-  val logger = Logger(this.getClass)
+  extends AbstractController(cc) with Logging {
 
   val metaGQLQ =
     gql"""
@@ -57,6 +54,7 @@ class MetaController @Inject()(implicit ec: ExecutionContext, backend: Backend, 
         case error: ErrorWithResolver => InternalServerError(error.resolveError)
       }
   }
+
   // example from here https://github.com/nemoo/play-slick3-example/blob/master/app/controllers/Application.scala
   def meta = Action.async { _ =>
     queryMeta
