+++ app/esecuele/Expression.scala
@@ -10,11 +10,13 @@ abstract class Expression extends Rep {
 
 case class RawExpression(raw: String) extends Expression {
   override lazy val rep: String = raw
+
   override def toString: String = rep
 }
 
 case object EmptyExpression extends Expression {
   override val raw: String = ""
   override lazy val rep: String = ""
+
   override def toString: String = rep
 }
