--- app/models/Entities.scala
@@ -1,15 +1,0 @@
-package models
-
-import models.entities.Violations.InputParameterCheckError
-import play.api.libs.json.Json
-import sangria.execution.{ExceptionHandler, HandledException, MaxQueryDepthReachedError}
-
-object Entities {
-  case object TooComplexQueryError extends Exception("Query is too expensive.")
-
-  lazy val exceptionHandler = ExceptionHandler {
-    case (_, error @ TooComplexQueryError) => HandledException(error.getMessage)
-    case (_, error @ MaxQueryDepthReachedError(_)) => HandledException(error.getMessage)
-    case (_, error @ InputParameterCheckError(_)) => HandledException(error.getMessage)
-  }
-}
