+++ app/models/entities/DDRelation.scala
@@ -1,6 +1,5 @@
 package models.entities
 
-import play.api.Logger
 import play.api.libs.json._
 import play.api.libs.json.Reads._
 import play.api.libs.functional.syntax._
