--- app/models/entities/Entity.scala
@@ -1,6 +1,0 @@
-package models.entities
-
-trait Entity {
-  def id: String
-}
-
